improve tests for remove_protocol function
